requirements.txt
python-telegram-bot
ccxt
pandas
pandas-ta
schedule